Implement takeOwnership()
