+++ contracts/assemblyscript-poc/contract/src/cosmwasm.ts
@@ -11,6 +11,34 @@ export class Region {
 }
 
 /**
+ * allocate reserves the given number of bytes in wasm memory and returns a pointer
+ * to a slice defining this data. This space is managed by the calling process
+ * and should be accompanied by a corresponding deallocate
+ */
+export function allocate(size: usize): usize {
+  const dataPtr = __alloc(size, idof<ArrayBuffer>());
+  __retain(dataPtr);
+
+  const region: Region = {
+    offset: dataPtr,
+    len: size,
+  };
+  const regionPtr = changetype<usize>(region);
+  __retain(regionPtr);
+  return regionPtr;
+}
+
+/**
+ * Expects a pointer to a Region created with allocate.
+ * It will free both the Region and the memory referenced by the Region.
+ */
+export function deallocate(regionPtr: usize): void {
+  const dataPtr = changetype<Region>(regionPtr).offset;
+  __release(regionPtr); // release Region
+  __release(dataPtr); // release ArrayBuffer
+}
+
+/**
  * Releases ownership of the data without destroying it.
  */
 export function releaseOwnership(data: Uint8Array): usize {
@@ -27,4 +55,18 @@ export function releaseOwnership(data: Uint8Array): usize {
   __retain(regionPtr);
 
   return regionPtr;
+}
+
+/**
+ * Takes ownership of the data at the given pointer
+ */
+export function takeOwnership(regionPtr: usize): Uint8Array {
+  const region = changetype<Region>(regionPtr);
+
+  const out = new Uint8Array(region.len);
+  // TODO: is this copy really necessary?
+  memory.copy(getDataPtr(out), region.offset, region.len);
+  deallocate(regionPtr);
+
+  return out;
 }
