+++ contracts/assemblyscript-poc/contract/src/index.ts
@@ -2,7 +2,9 @@
 import { JSONEncoder } from "assemblyscript-json";
 
 import * as contract from "./contract";
-import { releaseOwnership, Region } from "./cosmwasm";
+import { releaseOwnership } from "./cosmwasm";
+
+export { allocate, deallocate } from "./cosmwasm";
 
 function wrapSuccessData(data: Uint8Array): usize {
   const encoder = new JSONEncoder();
@@ -30,32 +32,4 @@ export function handle(_paramsPtr: usize, _messagePtr: usize): usize {
 
 export function query(_messagePtr: usize): usize {
   return wrapSuccessData(contract.query());
-}
-
-/**
- * allocate reserves the given number of bytes in wasm memory and returns a pointer
- * to a slice defining this data. This space is managed by the calling process
- * and should be accompanied by a corresponding deallocate
- */
-export function allocate(size: usize): usize {
-  const dataPtr = __alloc(size, idof<ArrayBuffer>());
-  __retain(dataPtr);
-
-  const region: Region = {
-    offset: dataPtr,
-    len: size,
-  };
-  const regionPtr = changetype<usize>(region);
-  __retain(regionPtr);
-  return regionPtr;
-}
-
-/**
- * Expects a pointer to a Region created with allocate.
- * It will free both the Region and the memory referenced by the Region.
- */
-export function deallocate(regionPtr: usize): void {
-  const dataPtr = changetype<Region>(regionPtr).offset;
-  __release(regionPtr); // release Region
-  __release(dataPtr); // release ArrayBuffer
 }
